Refer to getAsync method correctly when init is extension method.
Moved generateLibrary, along with the other related functions, into
a class called LibraryGenerator. This made it a little easier to
determine if the getAsync function should be referenced as get.getAsync
or simply getAsync. It also reduced the number of parameters that have
to be passed through to each method; instead each method can refer to
private fields.